requirements.txt
transformers==4.41.0
datasets==2.13.0
torch==2.3.1
pandas
fastapi
uvicorn
neo4j
huggingface-hub==0.23.0
tokenizers==0.19.1
langchain-huggingface==0.0.3
sentence-transformers==3.2.1
neontology
pydantic_settings
scrapy
python-dotenv
schedule
jina[standard]
jina[groundingapi]